Carletproject/requirements.txt
aniso8601==9.0.1
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dj-database-url==0.5.0
Django==3.1.7
django-extensions==3.1.1
django-filter==2.4.0
django-iban-field==0.8
django-phonenumber-field==5.0.0
djangorestframework==3.12.4
Flask==1.1.2
Flask-RESTful==0.3.8
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
phonenumbers==8.12.20
Pillow==8.1.2
Pygments==2.8.1
PySocks==1.7.1
pytz==2021.1
random-password-generator==2.1.0
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
webargs==8.0.0
Werkzeug==1.0.1
whitenoise==5.2.0
